not linking

+++ src/Macro/MacroManager.cpp
@@ -213,8 +213,8 @@ void MacroManager::update()
         curFrame = *i_s;
         double x1=(prvFrame - currentFrameCount)*hscale;
         double x2=(curFrame - currentFrameCount)*hscale;
-        Broodwar->drawLineScreen(x1,280-larvaCount*20,x2,280-larvaCount*20,Colors::Green);
-        Broodwar->drawLineScreen(x2,280-larvaCount*20,x2,280-(larvaCount+1)*20,Colors::Green);
+        Broodwar->drawLineScreen(static_cast<int>(x1),static_cast<int>(280-larvaCount*20),static_cast<int>(x2),static_cast<int>(280-larvaCount*20),Colors::Green);
+        Broodwar->drawLineScreen(static_cast<int>(x2),static_cast<int>(280-larvaCount*20),static_cast<int>(x2),static_cast<int>(280-(larvaCount+1)*20),Colors::Green);
         larvaCount++;
         i_s++;
       }
@@ -223,15 +223,15 @@ void MacroManager::update()
         curFrame = *i_u;
         double x1=(prvFrame - currentFrameCount)*hscale;
         double x2=(curFrame - currentFrameCount)*hscale;
-        Broodwar->drawLineScreen(x1,280-larvaCount*20,x2,280-larvaCount*20,Colors::Green);
-        Broodwar->drawLineScreen(x2,280-larvaCount*20,x2,280-(larvaCount-1)*20,Colors::Green);
+        Broodwar->drawLineScreen(static_cast<int>(x1),static_cast<int>(280-larvaCount*20),static_cast<int>(x2),static_cast<int>(280-larvaCount*20),Colors::Green);
+        Broodwar->drawLineScreen(static_cast<int>(x2),static_cast<int>(280-larvaCount*20),static_cast<int>(x2),static_cast<int>(280-(larvaCount-1)*20),Colors::Green);
         larvaCount--;
         i_u++;
       }
       prvFrame = curFrame;
     }
     double x1=(prvFrame - currentFrameCount)*hscale;
-    Broodwar->drawLineScreen(x1,280-larvaCount*20,640,280-larvaCount*20,Colors::Green);
+    Broodwar->drawLineScreen(static_cast<int>(x1),static_cast<int>(280-larvaCount*20),640,static_cast<int>(280-larvaCount*20),Colors::Green);
     Broodwar->drawLineScreen(0,280-0*20,640,280-0*20,Colors::Red);
   }
   }
